Add get_data to smhi parent and modify output
The time stamp used to output the Metobs data used to be one column with
Date and one column with time (UTC). Using pandas.to_datetime(), that
has been converted to one column for brevity according to comments in
discussion on issue on GitHub.

Also added umbrella function in SMHI class to get data.

Addresses #29.