Add interpolation from nearby station option
Adding an optional integer parameter "interpolate" with default value 0
which checks the immediate surroundings, in this solution implemented
using a fixed radius of 20km, of the chosen station for other stations
with the same parameter and uses the value from those parameters to fill
in time periods with blank data.

The assumption is that the median time between measurements is the
intended measurement interval for the parameter, as the header
information is somewhat ambiguously written making it difficult to
parse.

See also comment below.

Addresses #29.