Change d to nearby_stations, add default 0.
The name of the list specifying the nearest station from a given
position specified by lat/lon by the user is now called nearby_stations.
The default radius value to check for stations is set to 0, which
returns a list of exactly on station - the closest one.

Addresses #29.

(Because of merge issues, a reset has been made, rebuilding
functionality in this branch)